requirements.txt.txt
numpy
pandas
requests
datetime
plotly.express
time
hmac
hashlib
urllib
discord